feat(item): add mode switching functionality to tools
- Implement IItemMode interface for blaze and crystal tools
- Add mode switching logic to handle different tool behaviors
- Update tooltips and lang files for new mode information
- Refactor IInitEnchantItem to IItemEnchant for better naming convention

+++ src/main/java/committee/nova/mods/avaritia/api/utils/math/RayTracer.java
@@ -14,13 +14,12 @@ import net.minecraft.world.phys.Vec3;
 public class RayTracer {
 
     public static Vec3 getCorrectedHeadVec(Player player) {
-        return player.position().add(0.0D, (double) player.getEyeHeight(), 0.0D);
+        return player.position().add(0.0D, player.getEyeHeight(), 0.0D);
     }
 
     public static Vec3 getStartVec(Player player) {
         return getCorrectedHeadVec(player);
     }
-
 
     public static Vec3 getEndVec(Player player) {
         Vec3 headVec = getCorrectedHeadVec(player);
